+++ packages/test-base/src/commonTest/kotlin/io/realm/kotlin/test/common/RealmSetTests.kt
@@ -34,7 +34,6 @@ import io.realm.kotlin.test.common.utils.GenericTypeSafetyManager
 import io.realm.kotlin.test.common.utils.assertFailsWithMessage
 import io.realm.kotlin.test.platform.PlatformUtils
 import io.realm.kotlin.test.util.TypeDescriptor
-import io.realm.kotlin.types.ObjectId
 import io.realm.kotlin.types.RealmAny
 import io.realm.kotlin.types.RealmInstant
 import io.realm.kotlin.types.RealmObject
@@ -692,7 +691,6 @@ fun <T> getDataSetForCollectionClassifier(
     Decimal128::class -> if (nullable) NULLABLE_DECIMAL128_VALUES else DECIMAL128_VALUES
     String::class -> if (nullable) NULLABLE_STRING_VALUES else STRING_VALUES
     RealmInstant::class -> if (nullable) NULLABLE_TIMESTAMP_VALUES else TIMESTAMP_VALUES
-    ObjectId::class -> if (nullable) NULLABLE_OBJECT_ID_VALUES else OBJECT_ID_VALUES
     BsonObjectId::class -> if (nullable) NULLABLE_BSON_OBJECT_ID_VALUES else BSON_OBJECT_ID_VALUES
     RealmUUID::class -> if (nullable) NULLABLE_UUID_VALUES else UUID_VALUES
     ByteArray::class -> if (nullable) NULLABLE_BINARY_VALUES else BINARY_VALUES